requirements.txt
pymongo==3.7.2
python-dateutil==2.7.5
PyYAML==3.13
six==1.12.0
